% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_raw.R
\name{qc_raw}
\alias{qc_raw}
\title{Evaluates the quality control of raw files (FASTQ)}
\usage{
qc_raw(dir_fq = getwd(), threads = 4, dir_qc_raw = "output_qc_raw")
}
\arguments{
\item{dir_fq}{the path to the directory that contains the FASTQs files}

\item{threads}{number of threads}

\item{dir_qc_raw}{Path of the directory where the results of the function will be saved}
}
\value{
quality report
}
\description{
It assesses the quality control of FASTQ files before they are 
trimmed. It is the first function to be used.
}
\details{
This function needs the FastQC and MultiQC tools, as it will use these 
tools to generate the quality reports.
}
\author{
Jean Resende
}
