% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_trimming.R, R/trimming_trimmomatic.R
\name{qc_trimming}
\alias{qc_trimming}
\alias{trimming_trimmomatic}
\title{Quality control of trimmed files}
\usage{
qc_trimming(
  dir_fq = getwd(),
  threads = 4,
  dir_qc_trimming = "output_qc_trimming"
)

trimming_trimmomatic(input)
}
\arguments{
\item{dir_fq}{the path to the directory that contains the FASTQs files}

\item{threads}{number of threads}

\item{input}{directory path containing the FASTQs files to be trimmed.}

\item{dir_qc_trrimming}{Path of the directory where the results of the function will be saved}
}
\value{
quality report

trimmed FASTQ files
}
\description{
Evaluates the quality control of FASTQ files after they are
trimmed. It is the third function to use.

Trimming and filtering of readings. Second function used in the
process of obtaining transcript quantification.
}
\details{
This function needs the FastQC and MultiQC tools, as it will use these
tools to generate the quality reports.

This function makes use of the Trimmomatic tool.
}
\author{
Jean Resende
}
