% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon_index.R
\name{salmon_index}
\alias{salmon_index}
\title{Index construction}
\usage{
salmon_index(transc)
}
\arguments{
\item{transc}{Reference transcriptome path}
}
\value{
index
}
\description{
Construction of the reference transcriptome index.
This index will be used by the quant_salmon function.
}
\details{
This function makes use of the Salmon tool.
}
\author{
Jean Resende
}
