% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon_quant.R
\name{salmon_quant}
\alias{salmon_quant}
\title{Quantification of transcripts}
\usage{
salmon_quant(dir_fq, index)
}
\arguments{
\item{dir_fq}{Path to the directory with the FASTQs files}

\item{index}{Path to directory with transcriptome index}
}
\value{
transcript quantification
}
\description{
Quantification of transcripts based on the reference transcriptome.
This is the function that generates the quantification for each sample.
}
\details{
This function makes use of the Salmon tool.
}
\author{
Jean Resende
}
