% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming_trimmomatic.R
\name{trimming_trimmomatic}
\alias{trimming_trimmomatic}
\title{FASTQ Trimming}
\usage{
trimming_trimmomatic(input)
}
\arguments{
\item{input}{directory path containing the FASTQs files to be trimmed}

\value{
trimmed FASTQ files
}
\description{
Trimming and filtering of readings. Second function used in the
process of obtaining transcript quantification.
}
\details{
This function makes use of the Trimmomatic tool.
}
\author{
Jean Resende
}
